Have template functionality in a separate contract
