+++ packages/protocol/contracts/interfaces/ITemplateUtils.sol
@@ -1,0 +1,35 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.6;
+
+interface ITemplateUtils {
+    event CreatedTemplate(
+        bytes32 indexed templateId,
+        address airnode,
+        bytes32 endpointId,
+        bytes parameters
+    );
+
+    function createTemplate(
+        address airnode,
+        bytes32 endpointId,
+        bytes calldata parameters
+    ) external returns (bytes32 templateId);
+
+    function getTemplates(bytes32[] calldata templateIds)
+        external
+        view
+        returns (
+            address[] memory airnodes,
+            bytes32[] memory endpointIds,
+            bytes[] memory parameters
+        );
+
+    function templates(bytes32 templateId)
+        external
+        view
+        returns (
+            address airnode,
+            bytes32 endpointId,
+            bytes memory parameters
+        );
+}
